Add ability to send mail through SMTP
The only way to send mail was using sendgrid.com, which wasn't said in
the template config file and prevent people that don't want to pay for
mails to send mails.

This commit add an option to use a SMTP server. No SSL or authentication
options, but anyone is free to contribute :-)

The communication with the SMTP server is done with the help of cl-smtp,
licensed under the terms of the LLGPL (http://opensource.franz.com/preamble.html),
which seems to be compatible with Turtl licensing:
"The license allows developers and companies to use and integrate
software released under the LGPL into their own (even proprietary)
software"
(https://en.wikipedia.org/wiki/GNU_Lesser_General_Public_License)